fix py3 bugs
