+++ setup.py
@@ -1496,6 +1496,7 @@ def get_setup_parameters():
         'paddle.geometric',
         'paddle.geometric.message_passing',
         'paddle.geometric.sampling',
+        'paddle.ir',
     ]
 
     paddle_bins = ''
